Refactor roles pages with routing
