+++ src/pages/admin/RoleAddEdit.tsx
@@ -24,7 +24,7 @@ type Role = {
   }[];
 };
 
-function RoleForm() {
+function RoleAddEdit() {
   const { id } = useParams<{ id: string }>();
   const navigate = useNavigate();
   const { useQuery: useRoleQuery, useCreate, useUpdate } = useRoleRepository();
@@ -124,7 +124,7 @@ function RoleForm() {
           await supabase.from('role_permissions').insert(assignments);
         }
       }
-      navigate('/admin/roles');
+        navigate('/settings/administration/roles');
     } catch (error) {
       console.error('Error saving role:', error);
       if (error instanceof Error) {
@@ -167,7 +167,7 @@ function RoleForm() {
   return (
     <div className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8">
       <div className="mb-6">
-        <BackButton fallbackPath="/admin/roles" label="Back to Roles" />
+      <BackButton fallbackPath="/settings/administration/roles" label="Back to Roles" />
       </div>
 
       <div className="bg-white shadow overflow-hidden sm:rounded-lg">
@@ -285,7 +285,7 @@ function RoleForm() {
             <div className="mt-6 flex justify-end space-x-3">
               <button
                 type="button"
-                onClick={() => navigate('/admin/roles')}
+                onClick={() => navigate('/settings/administration/roles')}
                 className="inline-flex items-center px-4 py-2 border border-gray-300 shadow-sm text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-primary-500"
               >
                 Cancel
@@ -315,4 +315,4 @@ function RoleForm() {
   );
 }
 
-export default RoleForm;
+export default RoleAddEdit;
